changed to int

+++ src/main/java/com/changetemp/model/GoalTemperature.java
@@ -14,14 +14,14 @@ public class GoalTemperature {
     @Column(name = "T_Date")
     private LocalDate T_Date;
     @Column(name = "TValue")
-    private float TValue;
+    private int TValue;
     @Column(name = "T_Time")
     private LocalTime T_Time;
 
     public GoalTemperature() {
     }
 
-    public GoalTemperature(LocalDate t_Date, float TValue, LocalTime t_Time) {
+    public GoalTemperature(LocalDate t_Date, int TValue, LocalTime t_Time) {
         this.T_Date = t_Date;
         this.TValue = TValue;
         this.T_Time = t_Time;
@@ -43,11 +43,11 @@ public class GoalTemperature {
         this.T_Date = t_Date;
     }
 
-    public float getTValue() {
+    public int getTValue() {
         return TValue;
     }
 
-    public void setTValue(float TValue) {
+    public void setTValue(int TValue) {
         this.TValue = TValue;
     }
